Rename notifydialog as widgets/report_widget
