+++ client/widgets/report_widget.h
